requirements.txt
Flask==3.1.3
twilio==9.10.5
ollama==0.6.1
requests==2.33.1
python-dotenv==1.2.2
requests==2.33.1